Several migration fixes
- Changed `ULONG64` Windows specific typedef for the more standard
`unsigned long`.
- A few changes in the code because of the `QMatrix4x4` changes now
returning `float *` instead of `double *`.
- Solved issues with `gluUnProject()`.

+++ LivingCity/LC_GLWidget3D_Shadows.cpp
@@ -77,7 +77,7 @@ namespace LC {
 		//float znear=1500.0;
 
 		float f = 1.0f / tan (fov * (M_PI / 360.0));
-		double m[16]=
+        float m[16]=
 		{	f/aspect,	0,								0,									0,
 		0,			f,								0,						 			0,
 		0,			0,		(zfar+znear)/(znear-zfar),		(2.0f*zfar*znear)/(znear-zfar),
@@ -164,7 +164,7 @@ namespace LC {
 		float znear=1200.0f;// FIX
 
 		float f = 1.0f / tan (fov * (M_PI / 360.0));
-		double m[16]=
+        float m[16]=
 		{	f/aspect,	0,								0,									0,
 			0,			f,								0,						 			0,
 			0,			0,		(zfar+znear)/(znear-zfar),		(2.0f*zfar*znear)/(znear-zfar),
